+++ fnv.h
@@ -1,10 +1,10 @@
 #ifndef FNV_HEADER_FILE
 #define FNV_HEADER_FILE
 
-unsigned int fnv1a(const void *data, unsigned int numBytes);
-unsigned long long fnv1a64(
-    const void *data,
-    unsigned long long numBytes);
+#include <stdint.h>
+
+uint32_t fnv1a(const void *data, uint32_t numBytes);
+uint64_t fnv1a64(const void *data, uint64_t numBytes);
 
 #include "fnv.c"
 #endif
